+++ python-avd/pyavd/_eos_designs/structured_config/underlay/port_channel_interfaces.py
@@ -6,8 +6,8 @@ from __future__ import annotations
 from functools import cached_property
 from typing import TYPE_CHECKING
 
-from pyavd._eos_designs.interface_descriptions.models import InterfaceDescriptionData
 from pyavd._utils import get, short_esi_to_route_target
+from pyavd.api.interface_descriptions import InterfaceDescriptionData
 
 from .utils import UtilsMixin
